+++ easysession.el
@@ -959,6 +959,23 @@ The session name is displayed only when a session is actively loaded."
                 (setq-local trusted-content :all))))
           scratch))))
 
+(defun easysession--buffer-is-visible (buffer)
+  "Return non-nil if BUFFER is currently visible in the Emacs session.
+
+A buffer is considered visible if it is:
+
+- Displayed in any visible window (`get-buffer-window').
+- Associated with a visible tab in `tab-bar-mode' (if enabled).
+
+Returns nil if the buffer is not displayed in a window or tab."
+  (or
+   ;; Windows
+   (get-buffer-window buffer 'visible)
+   ;; Tab-bar windows
+   (and (bound-and-true-p tab-bar-mode)
+        (fboundp 'tab-bar-get-buffer-tab)
+        (tab-bar-get-buffer-tab buffer t nil))))
+
 ;;; Internal functions: handlers
 
 (defun easysession--restore-buffer-state (buffer buffer-info)
@@ -1391,6 +1408,44 @@ Returns a list:
      (easysession-undefine-save-handler ,key)))
 
 ;;; Autoloaded functions
+
+;;;###autoload
+(defun easysession-visible-buffer-list ()
+  "Return a list of all buffers considered visible in the current session.
+
+A buffer is included if it satisfies any of the following:
+- It is the *scratch* buffer (included as a special case).
+- It is currently displayed in a visible window.
+- It is associated with a visible tab in `tab-bar-mode', if enabled.
+
+The returned list contains live buffers only."
+  (let ((visible-buffers '()))
+    (dolist (buffer (buffer-list))
+      (when (and (buffer-live-p buffer)
+                 (or
+                  ;; Exception: The scratch buffer. (Useful for the
+                  ;; easysession-scratch extension.)
+                  (string= (buffer-name buffer) "*scratch*")
+
+                  ;; Buffers and indirect buffers
+                  (let ((base-buffer (buffer-base-buffer buffer)))
+                    (cond
+                     ;; Indirect buffers
+                     (base-buffer
+                      (and
+                       (buffer-live-p base-buffer)
+
+                       ;; Is the indirect buffer visible?
+                       (easysession--buffer-is-visible buffer)
+
+                       ;; Is the base buffer visible?
+                       (easysession--buffer-is-visible base-buffer)))
+
+                     ;; Normal buffers
+                     (t
+                      (easysession--buffer-is-visible buffer))))))
+        (push buffer visible-buffers)))
+    visible-buffers))
 
 ;;;###autoload
 (defun easysession-reset ()